Move two dirtools to a module
